refactor: 수치 컬럼 DECIMAL → DOUBLE 환원 및 BigDecimal 제거

+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/dashboard/DecisionRatioResult.java
@@ -3,13 +3,11 @@ package com.example.factoryguard.application.dto.dashboard;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Builder
 public class DecisionRatioResult {
 
     private final String decisionCode;
     private final Long count;
-    private final BigDecimal ratio;
+    private final Double ratio;
 }
