+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/dashboard/EquipmentScoreTrendResult.java
@@ -3,7 +3,6 @@ package com.example.factoryguard.application.dto.dashboard;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -13,5 +12,5 @@ public class EquipmentScoreTrendResult {
     private final Long targetId;
     private final String equipmentName;
     private final LocalDate date;
-    private final BigDecimal averageScore;
+    private final Double averageScore;
 }
