+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/result/RelatedResultResponse.java
@@ -3,7 +3,6 @@ package com.example.factoryguard.application.dto.result;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -12,7 +11,7 @@ public class RelatedResultResponse {
 
     private final Long resultId;
     private final LocalDateTime createdAt;
-    private final BigDecimal score;
+    private final Double score;
     private final String decisionCode;
     private final String location;
 }
