+++ backend-spring/src/main/java/com/example/factoryguard/domain/model/model/ModelVersion.java
@@ -4,7 +4,6 @@ import com.example.factoryguard.domain.model.vo.ModelDeployStatus;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,9 +14,9 @@ public class ModelVersion {
     private final Long modelId;
     private final Long fileId;
     private final String versionName;
-    private final BigDecimal accuracy;
-    private final BigDecimal precisionScore;
-    private final BigDecimal recallScore;
+    private final Double accuracy;
+    private final Double precisionScore;
+    private final Double recallScore;
     private final ModelDeployStatus deployStatus;
     private final Boolean isActive;
     private final LocalDateTime validatedAt;
